requirements.txt
python-telegram-bot>=13.7
requests>=2.28.0
python-dotenv>=0.19.0